NN-1796 Case note related pages
