+++ app/components/FormComponents/ButtonCancel/index.js
@@ -1,0 +1,3 @@
+import ButtonCancel from './ButtonCancel'
+
+export default ButtonCancel
