+++ app/components/FormComponents/TimePicker/TimePicker.test.js
@@ -1,9 +1,9 @@
 import React from 'react'
 import { shallow } from 'enzyme'
 import moment from 'moment'
-import { DATE_TIME_FORMAT_SPEC } from '../../../../containers/App/constants'
+import { DATE_TIME_FORMAT_SPEC } from '../../../containers/App/constants'
 
-import TimePicker from '../index'
+import TimePicker from '.'
 
 const setTime = (date, hours, minutes, seconds) => {
   date.hours(hours)
