+++ app/components/FormComponents/index.js
@@ -1,5 +1,8 @@
 import Submit from './Submit'
 import SubmissionError from './formComponents.theme'
 import TextArea from './TextArea'
+import Select from './Select'
+import TimePicker from './TimePicker'
+import ButtonCancel from './ButtonCancel'
 
-export { Submit, SubmissionError, TextArea }
+export { Submit, SubmissionError, TextArea, Select, TimePicker, ButtonCancel }
