+++ app/containers/Bookings/Details/AddCaseNote/AddCaseNote.test.js
@@ -1,5 +1,5 @@
 import { Map } from 'immutable'
-import { validate } from '../index'
+import { validate } from './AddCaseNote'
 
 describe('Add case note validation', () => {
   it('should check that type, subtype, comment and occurrenceDateTime have been entered', () => {
