--- app/containers/Bookings/Details/AddCaseNote/index.scss
@@ -1,28 +1,0 @@
-.add-case-note {
-
-   h1 {
-     margin-top: .5em;
-     margin-bottom: .5em;
-   }
-
-  .occurrence-date-time {
-    margin-bottom: 0.5em;
-  }
-
-  .form-control {
-    width: 100%;
-  }
-
-  .time-picker > .form-control {
-    width: 45% !important;
-    margin-right: 5px;
-  }
-
-  .rdtPicker {
-    width: 350px !important;
-  }
-
-  select {
-    height: 2em;
-  }
-}
