--- app/containers/Bookings/Details/CaseNotes/AmendCaseNote/index.scss
@@ -1,12 +1,0 @@
-.amend-case-note {
-
-  @include media-breakpoint-down(md) {
-    @include add-gutter-margin-top();
-  }
-
-  @include media-breakpoint-down(xs) {
-    button:first-child {
-      @include add-gutter-margin-bottom();
-    }
-  }
-}
