+++ app/containers/Bookings/Details/CaseNotes/filterForm.scss
@@ -61,15 +61,19 @@
     }
   }
 
-  .stack-type-subtype .form-group {
+  .stack-type-subtype label {
     @include large-inline-stack();
 
     @media(max-width: 991px) {
       @include small-inline-stack();
     }
+
+    select {
+      width: 100%;
+    }
   }
 
-  .stack-type-subtype .form-group:first-child {
+  .stack-type-subtype label:first-child {
      margin-left: 0;
   }
