+++ notm-specs/src/test/groovy/pages/AddCaseNotePage.groovy
@@ -16,14 +16,14 @@ class AddCaseNotePage extends Page {
     headingText { $("[data-qa=\'page-heading-text\']").text() }
     form { $('form') }
     type { $('select', name: 'typeValue') }
-    subType { $('select', name: 'subtypeValue') }
+    subType { $('select', name: 'subTypeValue') }
     textareaElement { $('textarea') }
     datePicker(required: false) { $('div.date-picker-component') } // click this to get picker
     days { $('td.rdtDay') } // days on picker, click to set date
     hours { $('select', name: 'hours') } // options 00, 01, 02 etc
     minutes { $('select', name: 'minutes') } // options 00, 05, 10 etc
     saveButton { $('button', type: 'submit') }
-    changeDateTime { $('button', name: 'change-date-time') }
+    changeDateTime { $("[data-qa=\'change-date-time\']") }
     typeSelectValue { $('form select')[0].value() }
     subTypeSelectValue { $('form select')[1].value() }
 
@@ -31,7 +31,12 @@ class AddCaseNotePage extends Page {
 
   def createNewCaseNote() {
     form.typeValue = 'CHAP'
+
     waitFor { form.find('option', value: 'FAITH').displayed }
+    form.subTypeValue = "FAITH"
+
+    // ensure that the javascript after selecting type has fired so our faith selection sticks
+    waitFor { form.subTypeValue == "FAITH" }
 
     textareaElement.module(Textarea).text = 'some text'
 
@@ -41,12 +46,6 @@ class AddCaseNotePage extends Page {
     days[0].click() // select 1st of this month for now
     form.hours = "07"
     form.minutes = "00"
-
-    // ensure that the javascript after selecting type has fired so our faith selection sticks
-    waitFor {
-      form.find('#subTypeValue').value('FAITH')
-      subTypeSelectValue == 'FAITH'
-    }
 
     saveButton.click()
   }
