+++ notm-specs/src/test/groovy/pages/CaseNotesPage.groovy
@@ -10,8 +10,8 @@ class CaseNotesPage extends Page {
     nextPageLink { $('#next-page') }
     previousPageLink { $('#previous-page') }
     caseNotes { $('.case-notes') }
-    typeSelect { $('#typeValue') }
-    subTypeSelect { $('#subTypeValue') }
+    typeSelect { $('select', name: 'typeValue') }
+    subTypeSelect { $('select', name: 'subTypeValue') }
     applyFiltersButton { $('form.filter-form button[type="submit"]') }
     resetFiltersButton { $('button.reset-filters-button') }
   }
