+++ package.json
@@ -51,6 +51,7 @@
     "@babel/runtime": "^7.3.4",
     "@contentful/rich-text-react-renderer": "^13.1.0",
     "@contentful/rich-text-types": "^13.1.0",
+    "@govuk-react/button": "^0.7.0",
     "@govuk-react/checkbox": "^0.6.0",
     "@govuk-react/constants": "^0.6.0",
     "@govuk-react/error-text": "^0.7.0",
